feat: reuse command hanlder. cleaning
