+++ ModularMonolith.User.Contracts/IUserService.cs
@@ -7,19 +7,4 @@ namespace ModularMonolith.User.Contracts
     {
         Task<UserDto> GetUserDetails(string userId);
     }
-
-    public class UserDto
-    {
-        public UserDto(string name, string surname, string email)
-        {
-            Name = name;
-            Surname = surname;
-            Email = email;
-        }
-        public string Name { get; }
-        public string Surname { get; }
-        public string Email { get; }
-
-        public string FullName => $"{Name} {Surname}";
-    }
 }
