+++ Modules/User/ModularMonolith.User.Application/Queries/GetUserDetails/GetUserDetailsQuery.cs
@@ -2,10 +2,11 @@
 using System.Collections.Generic;
 using System.Text;
 using MediatR;
+using ModularMonolith.User.Contracts;
 
 namespace ModularMonolith.User.Application.Queries.GetUserDetails
 {
-    public class GetUserDetailsQuery : IRequest<GetUserDetailsQueryResponse>
+    public class GetUserDetailsQuery : IRequest<UserDto>
     {
         public GetUserDetailsQuery(string userId)
         {
